dags/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astronomer-cosmos
apache-airflow-providers-postgres
dbt-core==1.7.13
dbt-postgres==1.7.13